fixed bunch of heirarchies & names to make it run
