+++ frontend/package.json
@@ -10,8 +10,8 @@
     "connected-react-router": "^6.8.0",
     "react": "^17.0.0",
     "react-dom": "^17.0.0",
-    "react-rating-stars-component": "^2.2.0",
     "react-meter-bar": "^1.0.2",
+    "react-rating-stars-component": "^2.2.0",
     "react-redux": "^7.2.1",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
