+++ frontend/src/App.js
@@ -2,8 +2,8 @@ import { BrowserRouter, Route, Redirect, Switch } from 'react-router-dom';
 import React from 'react';
 
 import DetailPage from './containers/DetailPage/DetailPage';
-import SignIn from "./containers/SignIn/SignIn";
 import MainPage from './containers/MainPage/MainPage'
+import SignIn from "./containers/SignIn/SignIn";
 import './App.css';
 
 function App() {
